+++ src/main/java/com/example/usermanagement/business/repository/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByNmId(Integer nmId);
+    Optional<User> findByNmId(Long nmId);
 
     Optional<User> findByDsEmail(String dsEmail);
 
-    boolean existsByEmail(String email, Integer userId);
+    boolean existsByEmail(String email, Long userId);
 }
